Move `ContextServerRegistry` into `context_servers`
