+++ crates/context_servers/src/context_servers.rs
@@ -6,13 +6,18 @@ use workspace::Workspace;
 pub mod client;
 pub mod manager;
 pub mod protocol;
+mod registry;
 pub mod types;
 
-actions!(context_servers, [Restart,]);
+pub use registry::*;
+
+actions!(context_servers, [Restart]);
 
 pub fn init(cx: &mut AppContext) {
     log::info!("initializing context server client");
     manager::init(cx);
+    ContextServerRegistry::register(cx);
+
     cx.observe_new_views(
         |workspace: &mut Workspace, _cx: &mut ViewContext<Workspace>| {
             workspace.register_action(restart_servers);
